Move aurora ice
