+++ Content/Tiles/Permafrost/AuroraIce.cs
@@ -1,4 +1,5 @@
 ﻿using StarlightRiver.Content.Abilities;
+using StarlightRiver.Content.Items.Permafrost;
 using StarlightRiver.Helpers;
 using System;
 using Terraria.GameContent;
@@ -165,132 +166,10 @@ namespace StarlightRiver.Content.Tiles.Permafrost
 		{
 			return Framing.GetTileSafely(x, y).TileType == TileID.IceBlock;
 		}
+
 		public string GetHint()
 		{
 			return "It fades away when you look at it...";
-		}
-	}
-
-	//TODO: Move all this to a more sane place, im really tired tonight and cant be assed to put braincells into organizing this. Thanks in advance future me.
-	class AuroraIceItem : QuickMaterial
-	{
-		public override string Texture => "StarlightRiver/Assets/Tiles/Permafrost/AuroraIceItem";
-
-		public AuroraIceItem() : base("Frozen Aurora Chunk", "A preserved piece of the night sky", 999, Item.sellPrice(0, 0, 5, 0), ItemRarityID.White) { }
-
-		public override bool PreDrawInInventory(SpriteBatch spriteBatch, Vector2 position, Rectangle frame, Color drawColor, Color itemColor, Vector2 origin, float scale)
-		{
-			Texture2D tex = TextureAssets.Item[Item.type].Value;
-			spriteBatch.Draw(tex, position, frame, itemColor, 0, origin, scale, 0, 0);
-
-			Color overColor = Color.White;
-			overColor.A = 0;
-
-			spriteBatch.Draw(tex, position, frame, overColor * 0.75f, 0, origin, scale, 0, 0);
-
-			return false;
-		}
-
-		public override bool PreDrawInWorld(SpriteBatch spriteBatch, Color lightColor, Color alphaColor, ref float rotation, ref float scale, int whoAmI)
-		{
-			Texture2D tex = TextureAssets.Item[Item.type].Value;
-			spriteBatch.Draw(tex, Item.position - Main.screenPosition, null, Item.color * 0.25f, rotation, tex.Size() / 2, scale, 0, 0);
-
-			Color overColor = Color.White;
-			overColor.A = 0;
-
-			spriteBatch.Draw(tex, Item.position - Main.screenPosition, null, overColor * 0.25f, rotation, tex.Size() / 2, scale, 0, 0);
-
-			return false;
-		}
-
-		public override void Update(ref float gravity, ref float maxFallSpeed)
-		{
-			float off = 0;
-			float time = Main.GameUpdateCount / 200f * 6.28f;
-
-			float sin2 = (float)Math.Sin(time + off * 0.2f * 0.2f);
-			float cos = (float)Math.Cos(time + off * 0.2f);
-			var color = new Color(100 * (1 + sin2) / 255f, 140 * (1 + cos) / 255f, 180 / 255f);
-
-			Item.color = color;
-		}
-
-		public override void UpdateInventory(Player player)
-		{
-			float off = 0;
-			float time = Main.GameUpdateCount / 200f * 6.28f;
-
-			float sin2 = (float)Math.Sin(time + off * 0.2f * 0.2f);
-			float cos = (float)Math.Cos(time + off * 0.2f);
-			var color = new Color(100 * (1 + sin2) / 255f, 140 * (1 + cos) / 255f, 180 / 255f);
-
-			Item.color = color;
-		}
-	}
-
-	class AuroraIceBar : QuickMaterial
-	{
-		public override string Texture => "StarlightRiver/Assets/Tiles/Permafrost/AuroraIceBar";
-
-		public AuroraIceBar() : base("Frozen Aurora Bar", "A preserved selection of the night sky", 99, Item.sellPrice(0, 0, 25, 0), ItemRarityID.Blue) { }
-
-		public override bool PreDrawInInventory(SpriteBatch spriteBatch, Vector2 position, Rectangle frame, Color drawColor, Color itemColor, Vector2 origin, float scale)
-		{
-			Texture2D tex = TextureAssets.Item[Item.type].Value;
-			spriteBatch.Draw(tex, position, frame, itemColor, 0, origin, scale, 0, 0);
-
-			Color overColor = Color.White;
-			overColor.A = 0;
-
-			spriteBatch.Draw(tex, position, frame, overColor * 0.75f, 0, origin, scale, 0, 0);
-
-			return false;
-		}
-
-		public override bool PreDrawInWorld(SpriteBatch spriteBatch, Color lightColor, Color alphaColor, ref float rotation, ref float scale, int whoAmI)
-		{
-			Texture2D tex = TextureAssets.Item[Item.type].Value;
-			spriteBatch.Draw(tex, Item.position - Main.screenPosition, null, Item.color * 0.25f, rotation, tex.Size() / 2, scale, 0, 0);
-
-			Color overColor = Color.White;
-			overColor.A = 0;
-
-			spriteBatch.Draw(tex, Item.position - Main.screenPosition, null, overColor * 0.25f, rotation, tex.Size() / 2, scale, 0, 0);
-
-			return false;
-		}
-
-		public override void Update(ref float gravity, ref float maxFallSpeed)
-		{
-			float off = 0;
-			float time = Main.GameUpdateCount / 200f * 6.28f;
-
-			float sin2 = (float)Math.Sin(time + off * 0.2f * 0.2f);
-			float cos = (float)Math.Cos(time + off * 0.2f);
-			var color = new Color(100 * (1 + sin2) / 255f, 140 * (1 + cos) / 255f, 180 / 255f);
-
-			Item.color = color;
-		}
-
-		public override void UpdateInventory(Player player)
-		{
-			float off = 0;
-			float time = Main.GameUpdateCount / 200f * 6.28f;
-
-			float sin2 = (float)Math.Sin(time + off * 0.2f * 0.2f);
-			float cos = (float)Math.Cos(time + off * 0.2f);
-			var color = new Color(100 * (1 + sin2) / 255f, 140 * (1 + cos) / 255f, 180 / 255f);
-
-			Item.color = color;
-		}
-
-		public override void AddRecipes()
-		{
-			Recipe recipe = CreateRecipe();
-			recipe.AddIngredient(ItemType<AuroraIceItem>(), 3);
-			recipe.AddTile(TileID.Furnaces);
-			recipe.Register();
 		}
 	}
 }
